rename index.ts to interfaces.ts
