+++ src/api/courses.ts
@@ -1,6 +1,6 @@
 import { firestore } from '../config/firebase';
 import { collection, getDocs, doc, setDoc, getDoc, updateDoc } from 'firebase/firestore';
-import { Course } from '../types';
+import { Course } from '../types/interfaces';
 
 export const getAllCourses = async (): Promise<Course[]> => {
   try {
@@ -17,11 +17,11 @@ export const addCourseToUser = async (userId: string, courseId: string): Promise
   try {
     const userDocRef = doc(firestore, 'users', userId);
     const userDoc = await getDoc(userDocRef);
-    
+
     if (userDoc.exists()) {
       const userData = userDoc.data();
       const userCourses = userData.courses || [];
-      
+
       if (!userCourses.includes(courseId)) {
         await updateDoc(userDocRef, {
           courses: [...userCourses, courseId]
@@ -40,20 +40,20 @@ export const getUserCourses = async (userId: string): Promise<Course[]> => {
   try {
     const userDocRef = doc(firestore, 'users', userId);
     const userDoc = await getDoc(userDocRef);
-    
+
     if (userDoc.exists()) {
       const userData = userDoc.data();
       const userCourses = userData.courses || [];
-      
+
       const coursesPromises = userCourses.map(async (courseId: string) => {
         const courseDocRef = doc(firestore, 'courses', courseId);
         const courseDoc = await getDoc(courseDocRef);
         return { id: courseDoc.id, ...courseDoc.data() } as Course;
       });
-      
+
       return Promise.all(coursesPromises);
     }
-    
+
     return [];
   } catch (error) {
     console.error('Error fetching user courses:', error);
